Improve sorting of items
Items have either a name, path or both. When resolving the attributes
the key makes up part of the string, which means that items are
sorted using the presence of a `name` attribute as a preference over
the actual contents of name/path.

This commit resolves this issue by adding a cononical_name method
that resolves to a sortable name without mentioning whether the
value came from the `name` or `path` attribute.